pyproject.toml
[project]
name = "retailvision-ai"
version = "0.1.0"
description = "Production-grade retail shelf intelligence system: YOLO-based product and gap detection with streaming analytics, forecasting, and GenAI insights."
readme = "README.md"
requires-python = ">=3.11,<3.13"
license = {text = "MIT"}
authors = [
    {name = "Maria Camila Gonzalez Nunez", email = "camila.gonzalez.ds@outlook.com"}
]
keywords = [
    "computer-vision",
    "object-detection",
    "yolo",
    "retail",
    "mlops",
    "streaming",
    "rag",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Programming Language :: Python :: 3.11",
]

# Production dependencies
dependencies = [
    # Core ML / CV
    "ultralytics>=8.4.37",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "opencv-python>=4.13.0",
    "pillow>=10.4.0",
    "numpy>=2.4.0",

    # Data handling
    "pandas>=2.3.0",
    "pyyaml>=6.0.0",
    "pydantic>=2.13.0",
    "scikit-learn>=1.8.0",

    # MLOps
    "mlflow>=3.11.0",
    "optuna>=4.8.0",

    # API layer
    "fastapi>=0.135.0",
    "uvicorn>=0.44.0",
    "python-multipart>=0.0.26",

    # Streaming & GCP
    "apache-beam[gcp]>=2.60.0,<2.72.0",
    "google-cloud-pubsub>=2.37.0",
    "google-cloud-bigquery>=3.41.0",
    "google-cloud-storage>=2.19.0",

    # Forecasting
    "statsforecast>=2.0.0",
    "lightgbm>=4.6.0",

    # GenAI / RAG
    "llama-index>=0.14.20",
    "llama-index-llms-gemini>=0.6.0",
    "llama-index-embeddings-gemini>=0.4.0",
    "google-genai>=1.73.0",

    # Dashboard
    "streamlit>=1.56.0",
    "plotly>=6.7.0",

    # Visualization (for notebooks)
    "matplotlib>=3.10.0",
    "seaborn>=0.13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=9.0.0",
    "pytest-cov>=7.1.0",
    "ruff>=0.6.0",
    "mypy>=1.10.0",
    "jupyter>=1.0.0",
    "ipykernel>=6.29.0",
]

[project.urls]
Homepage = "https://github.com/marianunez-data/retailvision-ai"
Repository = "https://github.com/marianunez-data/retailvision-ai"

# =============================================================================
# Build system
# =============================================================================
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

# =============================================================================
# Pytest configuration
# =============================================================================
[tool.pytest.ini_options]
minversion = "7.0"
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--strict-config",
    "--tb=short",
]
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "integration: integration tests requiring real dataset",
]

# =============================================================================
# Coverage configuration
# =============================================================================
[tool.coverage.run]
source = ["src"]
omit = [
    "tests/*",
    "*/__init__.py",
]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
]

# =============================================================================
# Ruff (linter + formatter)
# =============================================================================
[tool.ruff]
line-length = 88
target-version = "py311"

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

# =============================================================================
# MyPy (type checking)
# =============================================================================
[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false  # relaxed for notebooks
no_implicit_optional = true
check_untyped_defs = true
